Tu mensaje descriptivo aquí

+++ app/actions/addProducto/deteleteProducto.js
@@ -1,19 +1,26 @@
 "use server";
 
-import { PrismaClient } from '@prisma/client';
+
 import { revalidatePath } from 'next/cache';
+import prisma from "@/lib/prisma";
 
-const prisma = new PrismaClient();
 
-export default async function deleteProducto(formData) {
-  const id = parseInt(formData.get("id"));
+export default async function deleteProducto(id) {
+  console.log("El id recibido es:", id); // Añadir esta línea para verificar el id
+
+  if (typeof id !== 'number') {
+    throw new Error(`El id proporcionado no es un número válido: ${id}`);
+  }
 
   try {
     await prisma.producto.delete({
-      where: { id },
+      where: {
+        id: id,  // Asegúrate de que id es un número
+      },
     });
-    revalidatePath('/');
-  } catch (e) {
-    console.error(e);
+    console.log("Producto eliminado con éxito");
+  } catch (error) {
+    console.error("Error al eliminar el producto:", error);
+    throw error;
   }
 }
